+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # next (unreleased)
 
 * Escape underscores in Latex headers (Jendrik Seipp).
+* Remove `--dump-source` option (Jendrik Seipp).
 * Remove support for multiple input files (Jendrik Seipp).
 
 # 3.5 (2020-02-01)