+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Escape underscores in Latex headers (Jendrik Seipp).
 * Don't center Latex tables (Jendrik Seipp).
+* Use booktabs package for prettier Latex tables (Jendrik Seipp).
 * Remove `--dump-source` option (Jendrik Seipp).
 * Remove txt2tags version from output. (Jendrik Seipp).
 * Remove support for multiple input files (Jendrik Seipp).